requirements.txt
langchain
langchain-openai
langchain-community
langchain-chroma
chromadb
gitpython
PyGithub
streamlit
python-dotenv
tiktoken
